Bug 1669181 - Part 44: Split SafepointIndex and OsiIndex from "jit/JitFrames.h". r=jandem
`SafepointIndex` and `OsiIndex` only need a handful of dependencies, so split
them from the larger "jit/JitFrames.h" header.

Differential Revision: https://phabricator.services.mozilla.com/D92734
