+++ js/src/jit/arm/Bailouts-arm.cpp
@@ -7,6 +7,7 @@
 #include "jit/arm/Assembler-arm.h"
 #include "jit/Bailouts.h"
 #include "jit/JitRuntime.h"
+#include "jit/SafepointIndex.h"
 #include "vm/JSContext.h"
 #include "vm/Realm.h"
