+++ js/src/jit/arm64/Bailouts-arm64.cpp
@@ -5,6 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 #include "jit/Bailouts.h"
+#include "jit/SafepointIndex.h"
 
 #include "vm/JSScript-inl.h"
