+++ js/src/jit/shared/CodeGenerator-shared.h
@@ -18,6 +18,7 @@
 #include "jit/MacroAssembler.h"
 #include "jit/MIRGenerator.h"
 #include "jit/MIRGraph.h"
+#include "jit/SafepointIndex.h"
 #include "jit/Safepoints.h"
 #include "jit/Snapshots.h"
 #include "vm/TraceLoggingTypes.h"
